examples/aip-agent-quickstart/aip_agent_quickstart/agents/web_search_agent/web_search_agent/tools.py
"""Tools for the WebSearchAgent Agent."""

from langchain_core.tools import tool


@tool
def sample_tool(query: str) -> str:
    """A sample tool that processes a query and returns a string.

    This is a placeholder. Replace it with your actual agent tools.
    For example, it could be a tool to query a database, call an API,
    or perform a specific calculation related to web_search_agent.

    Args:
        query: The input string to process.

    Returns:
        A string representing the result of the tool's operation.
    """
    # TODO: Implement the actual logic for this tool
    return f"The WebSearchAgent Agent received your query: '{query}'. This is a sample tool response."


# Example of another tool:
# @tool
# def get_user_info(user_id: int) -> dict:
#     """Fetches user information from a database based on user_id."""
#     # Replace with actual database query or API call
#     if user_id == 1:
#         return {"user_id": 1, "name": "Alice", "email": "alice@example.com"}
#     return {"error": "User not found"}

# Add more tools as needed for your agent's capabilities.
