examples/aip-agent-quickstart/aip_agent_quickstart/agents/web_search_agent/pyproject.toml
[tool.poetry]
name = "web_search_agent"
version = "0.1.0"
description = "WebSearchAgent Agent - A2A Compliant"
authors = ["Your Name <you@example.com>"] # TODO: Replace with your details
readme = "README.md"
packages = [{include = "web_search_agent"}]
package-mode = false # Important for not trying to install the agent itself as a package

[tool.poetry.dependencies]
python = ">=3.13,<3.14"

# Core A2A and agent libraries
gllm-agents-binary = "0.2.6"

# LangChain (optional, but common for LLM agents)
langchain = "^0.3.0"
langchain_openai = "^0.3.0" # Or other LLM provider packages
langgraph = "^0.2.16"

# Other useful libraries (add as needed)
python-dotenv = "^1.0.0"
# pydantic = "^2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
